% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kro.inv.test.R
\name{elrt.alt}
\alias{elrt.alt}
\title{Simulating the extended LRT statistics under the alternative regime.}
\usage{
elrt.alt(
  n,
  p1,
  p2,
  center = TRUE,
  trans = TRUE,
  samp.num = 1000,
  samp.iter = 10,
  seed = 100
)
}
\description{
Simulating the extended LRT statistics under the alternative regime.
}
\examples{
p1=20; p2=16; n=50

}
\author{
Bongjung Sung
}
