% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utensils.R
\name{elrt.para}
\alias{elrt.para}
\title{Parameters of the transformation for the extended LRT statistic}
\usage{
elrt.para(p1, p2, n)
}
\arguments{
\item{p1}{the row dimension}

\item{p2}{the column dimension}

\item{n}{the sample size}
}
\value{
a parameter that is associated with transforming the extended LRT statistic based on the sample core.
}
\description{
Compute the parameters associated with transforming the extended LRT statistic based on the sample core.
}
\examples{

p1=20; p2=10; n=200
elrt.para(p1,p2,n)

}
\author{
Bongjung Sung
}
