% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/large.eig.test.R
\name{large.eig.alt}
\alias{large.eig.alt}
\title{Test statistic based on the largest eigenvalue of the sample core under the alternative regime}
\usage{
large.eig.alt(
  n,
  p1,
  p2,
  sigma,
  center = TRUE,
  trans = TRUE,
  sigma.known = FALSE,
  samp.num = 1000,
  iter = 10
)
}
\arguments{
\item{n}{the sample size.}

\item{p1}{the row dimension.}

\item{p2}{the column dimension.}

\item{sigma}{the population covariance matrix.}

\item{center}{logical, whether to center the data or not; TRUE by default.}

\item{trans}{logical,whether to transform the test statistic to approximate \code{TW}-law or not; TRUE by default.}

\item{sigma.known}{logical, whether to assume the known population covariance matrix or not; FALSE by default. If TRUE, the transformed test statistic is returned, regardless of \code{trans}.}

\item{samp.num}{the number of iterations for the Monte-Carlo simulation; 1000 by default.}

\item{iter}{the unit number at which to print the number of current iterations; 10 by default.}
}
\value{
the vector of \code{samp.num} Monte-Carlo simulated test statistics based on the largest eigenvalue of the sample core under \eqn{}.
}
\description{
Implement the Monte-Carlo simulations of the test statistic based on the largest eigenvalue of the sample core \eqn{\hat{C}} for Gaussian populations as benchmark cases under the alternative regime.
Here the sample core \eqn{\hat{C}} refers to the core component of the sample covariance matrix \eqn{S}.
The test statistic may be transformed to approximate the Tracy-Widom (\code{TW}) law if \code{trans}=TRUE, particularly under the local alternative regime (see Theorem 2 of Sung and Hoff (2025)). This transformation is based on the quantities depending only on \eqn{(n,p_1,p_2)}.
If the population covariance matrix of the mean is assumed to be known (\code{sigma.known}=TRUE), the transformation is done regardless of \code{trans} based on the Kronecker MLE (the separable component of \eqn{S}).
For the details, see (20) of Sung and Hoff (2025). Unless transformed, the function will return the values of \eqn{\lambda_1 (\hat{C})}.
Also, if the mean is assumed to be known, you may not center the data (\code{center=FALSE}). Otherwise, the data center should be centered.
}
\examples{
p1=20; p2=16; n=640

set.seed(100)
para.list=pi.rank2.core(p1,p2,lambda.gen=FALSE)

# local alternative
lambda1=1-(0.5/(320/2+0.5))
sigma1=pi.core(para.list,lambda0=lambda1)
test.stat1=large.eig.alt(n,p1,p2,sigma=sigma1,center=FALSE)
hist(test.stat1,freq=FALSE,breaks=25,xlab="Test Statistic",ylab="Density",main="Monte-Carlo Approximated Distribution")

# dense alternative
lambda2=1-(1.2/(320/2+1.2))
sigma2=pi.core(para.list,lambda0=lambda2)
test.stat2=large.eig.alt(n,p1,p2,sigma=sigma2,center=FALSE)
hist(test.stat2,freq=FALSE,breaks=25,xlab="Test Statistic",ylab="Density",main="Monte-Carlo Approximated Distribution")

}
\references{
Sung, B. and Hoff, P. (2025). Testing Separability of High-Dimensional Covariance Matrices.
arXiv preprint arXiv:2506.17463.
}
\author{
Bongjung Sung
}
