% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.isotropy.core.R
\name{pi.core}
\alias{pi.core}
\title{A covariance matrix with a partial-isotropy rank-r core.}
\usage{
pi.core(para.list, lambda0 = 0.5, root = "sym")
}
\arguments{
\item{para.list}{the list of parameters constituting the covariance matrix with a partial-isotropy rank-\eqn{r} core.}

\item{lambda0}{the pre-specified value of the non-spiked eigenvalue \code{lambda} in a partial-isotropy core. If it is not specified in \code{para.list} or does not belong to (0,1), it will be randomly generated; 0.5 by deafult.}

\item{root}{the choice of the square root of positive definite matrices; must be either \code{"sym"} (symmetric square root) or \code{"chol"} (Cholesky factor), \code{"sym"} by default.}
}
\value{
a \eqn{p1p2 \times p1p2} covariance matrix \eqn{\Sigma} with a partial-isotropy rank-\eqn{r} core.
The attribute \eqn{\lambda} of \eqn{\Sigma} denotes the value of the non-spiked eigenvalue.
}
\description{
Given a list of parameters constituting the covariance matrix with a partial-isotropy rank-\eqn{r} core, the function assembles the covariance matrix using these parameters.
}
\examples{

set.seed(100)
# generate a list of parameters for a covariance matrix of a partial-isotropy core.
p1=14; p2=10; r=3
para.list=pi.rank_r.core(p1,p2,r)
# assembles the covariance matrix using the above list of parameters.
pi.core(para.list)

}
\author{
Bongjung Sung
}
