% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial.isotropy.core.R
\name{pi.rank2.core}
\alias{pi.rank2.core}
\title{Partial-isotropy rank-2 core}
\usage{
pi.rank2.core(p1, p2, lambda.gen = TRUE)
}
\arguments{
\item{p1}{the row dimension.}

\item{p2}{the column dimension.}

\item{lambda.gen}{logical, whether to generate a non-spiked eigenvalue or not; TRUE by default.}
}
\value{
\code{pi.rank2.core} returns a list with the following elements:
\describe{
\item{K1}{the row covariance matrix of dimension \eqn{p1 \times p1};}
\item{K2}{the column covariance matrix of dimension \eqn{p2 \times p2};}
\item{A}{an array of factor matrices of a rank-2 core of dimension \eqn{p1 \times p2 \times 2};}
\item{lambda}{If \code{lambda.gen}=TRUE, the non-spiked eigenvalue \eqn{\lambda \in (0,1)}. Otherwise, \code{NULL}.}
}
}
\description{
Randomly generate a list of parameters for the covariance matrix  with a partial-isotropy rank-2 core for matrix-variate data.
}
\details{
If a core covariance matrix is of rank-2, the dimension \eqn{(p_1, p_2)} should satisfy one of the followings: \eqn{p_1 = p_2} or \eqn{p_1 - p_2 | p_1} \eqn{(p_1 > p_2)}, or \eqn{p_2 - p_1 | p_2} \eqn{(p_2 > p_1)}.
The covariance matrix \eqn{\Sigma} with a partial-isotropy rank-2 core takes the form of
\deqn{\Sigma = (K_2 \otimes K_1)^{1/2}((1- \lambda) A A^\top + \lambda I_{p_1 p_2})(K_2 \otimes K_1)^{1/2, \top}}
for \eqn{\lambda \in (0,1)}, positive definite matrices \eqn{K_1} and \eqn{K_2} of the dimensions \eqn{p_1 \times p_1} and \eqn{p_2 \times p_2}, respectively, and \eqn{A \in \mathbb{R}^{p_1p_2\times 2}} whose \eqn{i}th column is a vectorization of \eqn{p_1 \times p_2} matrix \eqn{A_i} such that \eqn{AA^\top} is a rank-2 core.
Here \eqn{M^{1/2}} denotes either symmetric square root or the Cholesky factor of a positive definite matrix \eqn{M}.
The separable and core components of \eqn{\Sigma}, denoted \eqn{K} and \eqn{C}, are
\deqn{K=K_2\otimes K_1,\quad C=(1- \lambda) A A^\top + \lambda I_{p_1 p_2}.}
For the exact formula, see Theorem 1 of Sung and Hoff (2025).
}
\examples{

set.seed(100)
# It must be that p1=p2 or p1-p2|p1 (p1>p2) or p2-p1|p2 (p2>p1).
p1=12; p2=10
pi.rank2.core(p1,p2)

}
\references{
Sung, B. and Hoff, P. (2025). Testing Separability of High-Dimensional Covariance Matrices.
arXiv preprint arXiv:2506.17463.
}
\author{
Bongjung Sung
}
