% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep.exp.test.R
\name{sep.exp.null}
\alias{sep.exp.null}
\title{Test statistic based on the separable expansion test under the null}
\usage{
sep.exp.null(
  n,
  p1,
  p2,
  center = TRUE,
  trans = TRUE,
  samp.num = 1000,
  iter = 10
)
}
\arguments{
\item{n}{the sample size.}

\item{p1}{the row dimension.}

\item{p2}{the column dimension.}

\item{center}{logical, whether to center the data or not; TRUE by default.}

\item{trans}{logical,whether to transform the test statistic; TRUE by default.}

\item{samp.num}{the number of iterations for the Monte-Carlo simulation; 1000 by default.}

\item{iter}{the unit number at which to print the number of current iterations; 10 by default.}
}
\value{
the vector of \code{samp.num} Monte-Carlo simulated null test statistics based on the separable expansion of the sample core.
}
\description{
Implement the Monte-Carlo simulations of the test statistic based on the separable expansion of the sample core \eqn{\hat{C}} for Gaussian populations as benchmark cases under the null hypothesis of separability.
Here the sample core \eqn{\hat{C}} refers to the core component of the sample covariance matrix \eqn{S}. The test statistic is given by
\deqn{ T(Y)=|| \hat{C} ||_F^2/p-1,}
 where \eqn{p = p_1 p_2} for row and column dimensions, \eqn{p_1} and \eqn{p_2}, respectively.
The test statistic may be transformed as \eqn{nT(Y)-p-1} if \code{trans}=TRUE.
For the details, see Section 3.3 and 5.1 of Sung and Hoff (2025). Also, if the mean is assumed to be known, you may not center the data (\code{center=FALSE}). Otherwise, the data center should be centered.
}
\examples{
p1=5; p2=3; n=60

set.seed(100)
sep.exp.null(n,p1,p2,center=FALSE,samp.num=20)

}
\references{
Sung, B. and Hoff, P. (2025). Testing Separability of High-Dimensional Covariance Matrices.
arXiv preprint arXiv:2506.17463.
}
\author{
Bongjung Sung
}
