% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kro.inv.test.R
\name{sep.exp.power}
\alias{sep.exp.power}
\title{Empirical power of the test based on the separable expansion}
\usage{
sep.exp.power(
  dat,
  center = TRUE,
  trans = TRUE,
  samp.num = 1000,
  samp.iter = 10,
  seed = 100
)
}
\description{
Empirical power of the test based on the separable expansion
}
\examples{
p1=20; p2=16; n=50

}
\author{
Bongjung Sung
}
