% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utensils.R
\name{sym.inv.root}
\alias{sym.inv.root}
\title{Inverse symmetric square root}
\usage{
sym.inv.root(cov)
}
\arguments{
\item{cov}{a positive definite matrix.}
}
\value{
the inverse of the symmetric square root of given a positive definite matrix \code{cov}.
}
\description{
Compute the inverse of the symmetric square root of a positive definite matrix.
}
\examples{

# generate a positive definite matrix
set.seed(100)
X=matrix(rnorm(4*10),ncol=4)
S=crossprod(X,X)/10
sym.inv.root(S)

}
\author{
Bongjung Sung
}
