% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utensils.R
\name{sym.root}
\alias{sym.root}
\title{Symmetric square root}
\usage{
sym.root(cov)
}
\arguments{
\item{cov}{a positive definite matrix.}
}
\value{
the symmetric square root of given a positive definite matrix \code{cov}.
}
\description{
Compute the symmetric square root of a positive definite matrix.
}
\examples{

# generate a positive definite matrix
set.seed(100)
X=matrix(rnorm(16*32),ncol=16)
S=crossprod(X,X)/32
sym.root(S)

}
\author{
Bongjung Sung
}
