% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utensils.R
\name{tw.para}
\alias{tw.para}
\title{Parameters of the transformation to obtain the Tracy-Widom law}
\usage{
tw.para(K1, K2, n)
}
\arguments{
\item{K1}{the \eqn{p1 \times p1} covariance matrix.}

\item{K2}{the \eqn{p2 \times p2} covariance matrix.}

\item{n}{the sample size.}
}
\value{
a vector of parameters that are associated with transforming the largest eigenvalue of the sample core to obtain \code{TW}-law.
}
\description{
Compute parameters associated with transforming the largest eigenvalue of the sample core to obtain \code{TW}-law.
}
\examples{

set.seed(100)
p1=3; p2=5; n=60; p=p1*p2
X=matrix(rnorm(p*n),ncol=p)
S=crossprod(X,X)/n
S.kcd=covKCD::covKCD(S,p1,p2)
K1=S.kcd$K1
K2=S.kcd$K2
tw.para(K1,K2,n)

}
\author{
Bongjung Sung
}
