% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elrt.R
\name{elrt.alt}
\alias{elrt.alt}
\title{Test statistic based on the extended LRT for sphericity testing under the alternative regime}
\usage{
elrt.alt(
  n,
  p1,
  p2,
  sigma,
  center = TRUE,
  trans = TRUE,
  samp.num = 1000,
  iter = 10
)
}
\arguments{
\item{n}{the sample size.}

\item{p1}{the row dimension.}

\item{p2}{the column dimension.}

\item{sigma}{the population covariance matrix.}

\item{center}{logical, whether to center the data or not; TRUE by default.}

\item{trans}{logical,whether to transform the test statistic; TRUE by default.}

\item{samp.num}{the number of iterations for the Monte-Carlo simulation; 1000 by default.}

\item{iter}{the unit number at which to print the number of current iterations; 10 by default.}
}
\value{
the vector of \code{samp.num} Monte-Carlo simulated test statistics based on the extended LRT, applied to the sample core, under the alternative.
}
\description{
Implement the Monte-Carlo simulations of the test statistic based on the the extended LRT, applied to the sample core \eqn{\hat{C}}, for Gaussian populations as benchmark cases under the alternative regime.
Here the sample core \eqn{\hat{C}} refers to the core component of the sample covariance matrix \eqn{S}. The test statistic may be transformed following Theorem 1--2 of Wang and Xu (2021) if \code{trans}=TRUE. This transformation is based on the quantities depending only on \eqn{(n,p_1,p_2)}.
For the details, see Section 3.2 and 5.1 of Sung and Hoff (2025).
The population covariance matrix should be specified in \code{sigma}. Also, if the mean is assumed to be known, you may not center the data (\code{center=FALSE}). Otherwise, the data center should be centered.
}
\examples{
p1=12; p2=10; r=4; n=200

set.seed(100)
para.list=pi.rank_r.core(p1,p2,r,lambda.gen=FALSE)
Sigma=pi.core(para.list,lambda0=0.95)

elrt.alt(n,p1,p2,Sigma,center=FALSE,samp.num=20)

}
\references{
Sung, B. and Hoff, P. (2025). Testing Separability of High-Dimensional Covariance Matrices.
arXiv preprint arXiv:2506.17463.

Wang, Z. and Xu, X. (2021). High-dimensional sphericity test by extended likelihood ratio.
Metrika 84:1169—-1212.
}
\author{
Bongjung Sung
}
