% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kro.inv.test.R
\name{large.eig.alt}
\alias{large.eig.alt}
\title{Simulating the largest eigenvalue test statistics under the alternative regime}
\usage{
large.eig.alt(
  n,
  p1,
  p2,
  sigma,
  center = TRUE,
  trans = TRUE,
  sigma.known = FALSE,
  samp.num = 1000,
  seed = 100
)
}
\description{
Simulating the largest eigenvalue test statistics under the alternative regime
}
\examples{
p1=20; p2=16; n=50

}
\author{
Bongjung Sung
}
