% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kro.inv.test.R
\name{large.eig.null}
\alias{large.eig.null}
\title{Simulating the largest eigenvalue test statistics under the null hypothesis of separability}
\usage{
large.eig.null(
  n,
  p1,
  p2,
  center = TRUE,
  trans = TRUE,
  sigma.known = FALSE,
  samp.num = 1000,
  samp.iter = 10,
  seed = 100
)
}
\description{
Simulating the largest eigenvalue test statistics under the null hypothesis of separability
}
\examples{
p1=20; p2=16; n=50

}
\author{
Bongjung Sung
}
