% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kro.inv.test.R
\name{sep.exp.null}
\alias{sep.exp.null}
\title{Simulating the separable expansion test statistics under the null hypothesis}
\usage{
sep.exp.null(
  n,
  p1,
  p2,
  center = TRUE,
  trans = TRUE,
  samp.num = 1000,
  samp.iter = 10,
  seed = 100
)
}
\description{
Simulating the separable expansion test statistics under the null hypothesis
}
\examples{
p1=20; p2=16; n=50

}
\author{
Bongjung Sung
}
