% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utensils.R
\name{tw.para}
\alias{tw.para}
\title{Parameters of the transformation to obtain the Tracy-Widom law}
\usage{
tw.para(K1, K2, n)
}
\arguments{
\item{K1}{the \eqn{p1 \times p1} covariance matrix.}

\item{K2}{the \eqn{p2 \times p2} covariance matrix.}

\item{n}{the sample size.}
}
\value{
a vector of parameters that are associated with transforming the largest eigenvalue of the sample core to obtain \code{TW}-law.
}
\description{
Compute parameters associated with transforming the largest eigenvalue of the sample core to obtain \code{TW}-law.
}
\examples{

set.seed(100)
p1=20; p2=12; n1=40; n2=24
X1=matrix(rnorm(p1*n1),ncol=p1)
X2=matrix(rnorm(p2*n2),ncol=p2)
K1=crossprod(X1,X1)/n1
K2=crossprod(X2,X2)/n2
n=400
tw.para(K1,K2,n)

}
\author{
Bongjung Sung
}
